# app/backend/openapi.yaml
openapi: 3.0.1
info:
  title: Gigator API
  version: 0.1.0
  description: Minimal OpenAPI skeleton for core Gigator backend endpoints (opportunities, threads, compose, auth, AI).
servers:
  - url: http://localhost:3000
    description: Local dev
paths:
  /api/venues:
    get:
      summary: List venues
      responses:
        '200':
          description: Array of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
    post:
      summary: Create venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - city
              properties:
                name:
                  type: string
                city:
                  type: string
                notes:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Created venue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
  /api/venues/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get venue by id
      responses:
        '200':
          description: Venue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "not_found"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
    patch:
      summary: Update venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                city:
                  type: string
                notes:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "not_found"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      description: Generic error response payload.
      required:
        - error
      properties:
        error:
          type: string
          description: Stable error identifier.
        message:
          type: string
          description: Human-readable error message (safe to expose).
...
    Venue:
      type: object
      required:
        - id
        - name
        - city
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
...